# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["users"]
      }
    },
...
    "/admin/users": {
      "get": {
        "operationId": "AdminUsersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-users"]
      }
    },
    "/admin/users/{id}/role": {
      "patch": {
        "operationId": "AdminUsersController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-users"]
      }
    },
    "/admin/users/{id}/ban": {
      "patch": {
        "operationId": "AdminUsersController_ban",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-users"]
      }
    },
    "/admin/users/{id}/unban": {
      "patch": {
        "operationId": "AdminUsersController_unban",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-users"]
      }
    },
    "/admin/users/{id}": {
      "delete": {
        "operationId": "AdminUsersController_anonymize",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["admin-users"]
      }
    },
...
  },
  "info": {
    "title": "klopilot.ch API",
    "description": "Community-Plattform für Toilettenbewertungen.\n\n**Public API (v1):** Endpunkte unter `/v1/` benötigen einen `X-Api-Key`-Header.\nKey erstellen: `POST /api-keys` (JWT-Auth erforderlich).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "klo_refresh"
      },
      "X-Api-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
...
  }
}